+++ src/components/designSystem/Filters/filtersElements/FiltersItemCreditNoteCreditStatus.tsx
@@ -2,7 +2,7 @@ import { MultipleComboBox } from '~/components/form'
 import { CreditNoteCreditStatusEnum } from '~/generated/graphql'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
 
-import { FiltersFormValues } from '../FiltersPanelPopper'
+import { FiltersFormValues } from '../types'
 
 type FiltersItemCreditNoteCreditStatusProps = {
   value: FiltersFormValues['filters'][0]['value']
