+++ src/components/designSystem/Filters/filtersElements/FiltersItemInvoiceType.tsx
@@ -2,7 +2,7 @@ import { MultipleComboBox } from '~/components/form'
 import { InvoiceTypeEnum } from '~/generated/graphql'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
 
-import { FiltersFormValues } from '../FiltersPanelPopper'
+import { FiltersFormValues } from '../types'
 
 type FiltersItemInvoiceTypeProps = {
   value: FiltersFormValues['filters'][0]['value']
